refactored tests to separate the tests, which listen on the processor event bus address
