+++ src/test/java/org/swisspush/redisques/AbstractTestCase.java
@@ -31,10 +31,8 @@ import static org.swisspush.redisques.util.RedisquesAPI.*;
 @RunWith(VertxUnitRunner.class)
 public abstract class AbstractTestCase {
 
-    public static final int NUM_QUEUES = 10;
-    public static final String TIMESTAMP = "timestamp";
-    public static final String QUEUES_PREFIX = "redisques:queues:";
-    public static final String REDISQUES_LOCKS = "redisques:locks";
+
+
 
     @Rule
     public Timeout rule = Timeout.seconds(5);
@@ -93,43 +91,7 @@ public abstract class AbstractTestCase {
 
     private static void initProcessor(EventBus eventBus){
 
-        final Map<String, MessageDigest> signatures = new HashMap<>();
 
-        eventBus.consumer("processor-address", new Handler<Message<JsonObject>>() {
-            public void handle(final Message<JsonObject> message) {
-                final String queue = message.body().getString("queue");
-                final String payload = message.body().getString("payload");
-
-                if ("STOP".equals(payload)) {
-                    message.reply(new JsonObject() {
-                        {
-                            put("status", "ok");
-                        }
-                    });
-                    eventBus.send("digest-" + queue, DatatypeConverter.printBase64Binary(signatures.get(queue).digest()));
-                } else {
-                    MessageDigest signature = signatures.get(queue);
-                    if (signature == null) {
-                        try {
-                            signature = MessageDigest.getInstance("MD5");
-                            signatures.put(queue, signature);
-                        } catch (NoSuchAlgorithmException e) {
-                            throw new RuntimeException();
-                        }
-                    }
-                    signature.update(payload.getBytes());
-                }
-
-                vertx.setTimer(new Random().nextLong() % 1 + 1, event -> {
-                    log.debug("Processed message " + payload);
-                    message.reply(new JsonObject() {
-                        {
-                            put("status", "ok");
-                        }
-                    });
-                });
-            }
-        });
     }
 
     @Before
@@ -137,70 +99,7 @@ public abstract class AbstractTestCase {
         flushAll();
     }
 
-    int numMessages = 5;
-    AtomicInteger finished = new AtomicInteger();
 
-    /**
-     * Sender Class
-     */
-    class Sender {
-        final String queue;
-        int messageCount;
-        MessageDigest signature;
-        TestContext context;
-        Async async;
 
-        Sender(TestContext context, Async async, final String queue) {
-            this.context = context;
-            this.async = async;
-            this.queue = queue;
-            try {
-                signature = MessageDigest.getInstance("MD5");
-            } catch (NoSuchAlgorithmException e) {
-                throw new RuntimeException(e);
-            }
 
-            vertx.eventBus().consumer("digest-" + queue, new Handler<Message<String>>() {
-                @Override
-                public void handle(Message<String> event) {
-                    log.info("Received signature for " + queue + ": " + event.body());
-                    context.assertEquals(event.body(), DatatypeConverter.printBase64Binary(signature.digest()), "Signatures differ");
-                    if (finished.incrementAndGet() == NUM_QUEUES) {
-                        async.complete();
-                    }
-                }
-            });
-        }
-
-        void send(final String m) {
-            if (messageCount < numMessages) {
-                final String message;
-                if(m==null) {
-                    message = Double.toString(Math.random());
-                } else {
-                    message = m;
-                }
-                signature.update(message.getBytes());
-                vertx.eventBus().send("redisques", buildEnqueueOperation(queue, message), new Handler<AsyncResult<Message<JsonObject>>>() {
-                    @Override
-                    public void handle(AsyncResult<Message<JsonObject>> event) {
-                        if(event.result().body().getString(STATUS).equals(OK)) {
-                            send(null);
-                        } else {
-                            log.error("ERROR sending "+message+" to "+queue);
-                            send(message);
-                        }
-                    }
-                });
-                messageCount++;
-            } else {
-                vertx.eventBus().send("redisques", buildEnqueueOperation(queue, "STOP"), new Handler<AsyncResult<Message<JsonObject>>>() {
-                    @Override
-                    public void handle(AsyncResult<Message<JsonObject>> reply) {
-                        context.assertEquals(OK, reply.result().body().getString(STATUS));
-                    }
-                });
-            }
-        }
-    }
 }
